Translations actually works, functions renamed
Padding ones at the end works but padding ones at the beginning previously didn't. I'm not going to question it.

In other news, some functions were moved around to make more sense. Gradient is not a part of regression and generate_training_set is a static method of TrainingSet.
Speaking of TrainingSet, Training_Set felt wrong so I read the PEP guidelines and it now conforms to python standard.